+++ jsonutil/jsonutil_test.go
@@ -11,6 +11,7 @@ import (
 	"errors"
 	"io/ioutil"
 	"os"
+	"runtime"
 	"testing"
 
 	. "github.com/essentialkaos/check"
@@ -77,6 +78,7 @@ func (s *JSONSuite) TestDecoding(c *C) {
 	err = Read(s.TmpDir+"/file-not-exists.json", &TestStruct{})
 
 	c.Assert(err, NotNil)
+	c.Assert(err, ErrorMatches, `open .*/file-not-exists.json: no such file or directory`)
 
 	err = Read(s.TmpDir+"/file1.json", testStruct)
 
@@ -109,6 +111,12 @@ func (s *JSONSuite) TestEncoding(c *C) {
 	err = Write("/test.json", testStruct)
 
 	c.Assert(err, NotNil)
+
+	if runtime.GOOS == "darwin" {
+		c.Assert(err, ErrorMatches, `open /test.json: read-only file system`)
+	} else {
+		c.Assert(err, ErrorMatches, `open /test.json: permission denied`)
+	}
 
 	err = Write(jsonFile, map[float64]int{3.14: 123})