+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 * `[directio]` Improved tests
 * `[fsutil]` Improved tests
 * `[initsystem]` Improved tests
+* `[jsonutil]` Improved tests
 * `[knf/validators/fs]` Improved tests
 * `[knf/validators/network]` Improved tests
 * `[knf/validators/regexp]` Improved tests